Fixed bug where after pjax request, adding new row caused double insert.

+++ src/DynamicFormWidget.php
@@ -201,20 +201,22 @@ class DynamicFormWidget extends \yii\base\Widget
     {
         DynamicFormAsset::register($view);
 
-        // add a click handler for the clone button
-        $js = 'jQuery(document).on("click", "#'. $this->formId . ' '. $this->insertButton . '", function(e) {'. "\n";
-        $js .= "    e.preventDefault();\n";
-        $js .= '    jQuery(".' .  $this->widgetContainer . '").triggerHandler("beforeInsert", [jQuery(this)]);' . "\n";
-        $js .= '    jQuery(".' .  $this->widgetContainer . '").yiiDynamicForm("addItem", '. $this->_hashVar . ", e, jQuery(this));\n";
-        $js .= "});\n";
-        $view->registerJs($js, $view::POS_READY);
+        if (!Yii::$app->request->isPjax) {
+            // add a click handler for the clone button
+            $js = 'jQuery(document).on("click", "#'. $this->formId . ' '. $this->insertButton . '", function(e) {'. "\n";
+            $js .= "    e.preventDefault();\n";
+            $js .= '    jQuery(".' .  $this->widgetContainer . '").triggerHandler("beforeInsert", [jQuery(this)]);' . "\n";
+            $js .= '    jQuery(".' .  $this->widgetContainer . '").yiiDynamicForm("addItem", '. $this->_hashVar . ", e, jQuery(this));\n";
+            $js .= "});\n";
+            $view->registerJs($js, $view::POS_READY);
 
-        // add a click handler for the remove button
-        $js = 'jQuery(document).on("click", "#'. $this->formId . ' '. $this->deleteButton . '", function(e) {'. "\n";
-        $js .= "    e.preventDefault();\n";
-        $js .= '    jQuery(".' .  $this->widgetContainer . '").yiiDynamicForm("deleteItem", '. $this->_hashVar . ", e, jQuery(this));\n";
-        $js .= "});\n";
-        $view->registerJs($js, $view::POS_READY);
+            // add a click handler for the remove button
+            $js = 'jQuery(document).on("click", "#'. $this->formId . ' '. $this->deleteButton . '", function(e) {'. "\n";
+            $js .= "    e.preventDefault();\n";
+            $js .= '    jQuery(".' .  $this->widgetContainer . '").yiiDynamicForm("deleteItem", '. $this->_hashVar . ", e, jQuery(this));\n";
+            $js .= "});\n";
+            $view->registerJs($js, $view::POS_READY);
+        }
 
         $js = 'jQuery("#' . $this->formId . '").yiiDynamicForm(' . $this->_hashVar .');' . "\n";
         $view->registerJs($js, $view::POS_LOAD);
